Issue #1 ------- Fixed bad password message
Issue #2 and #12
-------------
Possible fix.  Added code to turn of the spinner

Issue #9
-------
SVG files for the Knightly orders now local to the app.  Images for Peerages will still have to be loaded from the server

Layout of new user signup form completed, with validation and check that the email address is not in use.  Form currently does not actually submit the information to Medusa

Added splash/start-up screen

Add some build automation